// types/graphql-types.d.ts
export type Maybe<T> = T | null
export type Exact<T extends { [key: string]: any }> = { [K in keyof T]: T[K] }
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  /** The `ID` scalar type represents a unique identifier, often used to refetch an object or as key for a cache. The ID type appears in a JSON response as a String; however, it is not intended to be human-readable. When expected as an input type, any string (such as `"4"`) or integer (such as `4`) input value will be accepted as an ID. */
  ID: string
  /** The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text. */
  String: string
  /** The `Boolean` scalar type represents `true` or `false`. */
  Boolean: boolean
  /** The `Int` scalar type represents non-fractional signed whole numeric values. Int can represent values between -(2^31) and 2^31 - 1. */
  Int: number
  /** The `Float` scalar type represents signed double-precision fractional values as specified by [IEEE 754](https://en.wikipedia.org/wiki/IEEE_floating_point). */
  Float: number
  /** A date string, such as 2007-12-03, compliant with the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  Date: any
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any
}
// ...
export type Node = {
  id: Scalars["ID"]
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
}

export type Internal = {
  content?: Maybe<Scalars["String"]>
  contentDigest: Scalars["String"]
  description?: Maybe<Scalars["String"]>
  fieldOwners?: Maybe<Array<Maybe<Scalars["String"]>>>
  ignoreType?: Maybe<Scalars["Boolean"]>
  mediaType?: Maybe<Scalars["String"]>
  owner: Scalars["String"]
  type: Scalars["String"]
}
// ...
export type SiteSiteMetadata = {
  title?: Maybe<Scalars["String"]>
  description?: Maybe<Scalars["String"]>
  author?: Maybe<SiteSiteMetadataAuthor>
  siteUrl?: Maybe<Scalars["String"]>
  social?: Maybe<SiteSiteMetadataSocial>
}

export type SiteSiteMetadataAuthor = {
  name?: Maybe<Scalars["String"]>
  summary?: Maybe<Scalars["String"]>
}

export type SiteSiteMetadataSocial = {
  twitter?: Maybe<Scalars["String"]>
  github?: Maybe<Scalars["String"]>
}
// ...
export type MarkdownHeading = {
  id?: Maybe<Scalars["String"]>
  value?: Maybe<Scalars["String"]>
  depth?: Maybe<Scalars["Int"]>
}

export type MarkdownHeadingLevels = "h1" | "h2" | "h3" | "h4" | "h5" | "h6"

export type MarkdownExcerptFormats = "PLAIN" | "HTML" | "MARKDOWN"

export type MarkdownWordCount = {
  paragraphs?: Maybe<Scalars["Int"]>
  sentences?: Maybe<Scalars["Int"]>
  words?: Maybe<Scalars["Int"]>
}

export type MarkdownRemark = Node & {
  id: Scalars["ID"]
  frontmatter?: Maybe<MarkdownRemarkFrontmatter>
  excerpt?: Maybe<Scalars["String"]>
  rawMarkdownBody?: Maybe<Scalars["String"]>
  fileAbsolutePath?: Maybe<Scalars["String"]>
  fields?: Maybe<MarkdownRemarkFields>
  html?: Maybe<Scalars["String"]>
  htmlAst?: Maybe<Scalars["JSON"]>
  excerptAst?: Maybe<Scalars["JSON"]>
  headings?: Maybe<Array<Maybe<MarkdownHeading>>>
  timeToRead?: Maybe<Scalars["Int"]>
  tableOfContents?: Maybe<Scalars["String"]>
  wordCount?: Maybe<MarkdownWordCount>
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
}
// ...
export type MarkdownRemarkFrontmatter = {
  title?: Maybe<Scalars["String"]>
  date?: Maybe<Scalars["Date"]>
  description?: Maybe<Scalars["String"]>
  tags?: Maybe<Array<Maybe<Scalars["String"]>>>
}

export type MarkdownRemarkFrontmatterDateArgs = {
  formatString?: Maybe<Scalars["String"]>
  fromNow?: Maybe<Scalars["Boolean"]>
  difference?: Maybe<Scalars["String"]>
  locale?: Maybe<Scalars["String"]>
}

export type MarkdownRemarkFields = {
  slug?: Maybe<Scalars["String"]>
}
// ...
export type BlogPostBySlugQuery = {
  site?: Maybe<{ siteMetadata?: Maybe<Pick<SiteSiteMetadata, "title">> }>
  markdownRemark?: Maybe<
    Pick<MarkdownRemark, "id" | "excerpt" | "html"> & {
      frontmatter?: Maybe<Pick<MarkdownRemarkFrontmatter, "title" | "date" | "description" | "tags">>
    }
  >
}